feat: Global network with direct resolution in hosts_file
Allow to define a network as "global network" and an Ansible group related to it. Then, in /etc/hosts files, hosts in this Ansible group will have their direct hostname resolution on this sepcific network instead of being on their j2_node_main_network. This is very useful for example for multi-iceberg systems with an interconnect network, so computational nodes can contact themself directly on the interconnect network to initiate MPI communications, rempving the need of routes.