Feat: make yum repositories more configurable
Allow to modify some settings for the repo, be it a system repository or
an external repository.

List of settings: name, baseurl, enabled, gpgcheck, gpgkey and proxy.

System repositories, internal to the cluster, can use the same
configuration syntax than previously:

  repositories:
    - bluebanquise
    - os

It can be expanded to change settings for a repo with a dict, using
YAML's flow styles or block styles:

  repositories:
    - bluebanquise
    - { name: 'os', gpgcheck: 1 }  # Flow style
    - name: epel                   # Block style
      gpgcheck: 1

Additional repostories can either be added to the repositories.yml or
external.yml configuration files, to logically segregate internal and
external repositories.

  external_repositories:
    - name: epel
      baseurl: 'https://dl.fedoraproject.org/pub/epel/$releasever/$basearch/'
      proxy: 'https://proxy:8080'