Feature: alternate next-server for HTTP/iPXE
By default, the pxe_ip defines the next-server to use for both the TFTP
and HTTP/iPXE services which muse be hosted on the same server.

It is now possible to use a different next-server for HTTP/iPXE by
defining a new service parameter pxe_http_ip, thus the HTTP server for
ipxe scripts can be run on a separate server then the TFTP service.

The PXE chain remains unchanged: the first next-server obtained at boot
time is pxe_ip and it allows to load the iPXE bluebanquise ROM through
TFTP. This ROM will send a new request to the DHCP to get the
next-server to use for the HTTP/iPXE service. If pxe_http_ip is set,
then this is the IP address that will be used as next-server. Otherwise,
the default will be the pxe_ip.

Solve #60.