Update main.yml
When deploying services like dhcp, tftp, http, dns etc. inside containers running on various bare metal nodes, we used the ansible "chroot" connector to generate the configuration files on a shared file system mounted on all nodes.
We want to use the standard roles (dhcp, dns, pxe_stack ...) for that purpose, but we do NOT want to start the services on the bare metal management node. So we use the template tag , but we need also for pxe_stack role to create the output directories and to copy the bootswitch cgi script. So, a directory and a script tag are needed for that.