feat: Disable CentOS distro repositories
This feature allows to disable the CentOS repositories enabled by
default after the installation of the OS. It can be enabled or disabled
with the boolean parameter remove_distro_repositories.

If the parameter is not defined, the repositories are untouched to keep
current behavior. In the examples, the parameter defaults to True
because it is good practice to set up a local mirror when we have a lot
of systems.