feat: configure several DNS servers in the clients
Allow to configure a list of several name servers in the dns_ip
parameter of a network:

  dns_ip:
    - 10.11.0.1
    - 10.11.0.2

Update the dns_client role to create as much nameserver entries in
/etc/resolv.conf.

Update the dhcp_server and advanced_dhcp_server roles to add all the
dns_ip entries to domain-name-servers.

Enable jinja2's lstrip_blocks to indent the j2 templates without
indenting the destination file.