Trailing whitespace in dhcpd.subnet.conf.j2
Hello,
When we define several networks (ice1-2, ice1-3, ...) , the dhcpd.subnet.conf.j2 template generate many trailing whitespace in DHPC configuration files.
This is not a problem to run the DHCP service, but the configuration files generate is too big and not very clean.

For fix this issue, please merge my patch.

Regards,