feat: tune NM_CONTROLLED in nic role
Add a boolean `nic_nm_controlled` to tune the NM_CONTROLLED parameter in
the ifcfg files. The value defaults to `yes`.

To disable NM_CONTROLLED, update the inventory with:

  ```yaml
  nic_nm_controlled: no
  ```

This is a global parameter. For a single host, it is not possible to
control one network interface with NetworkManager and others without.

Closes: #131