fix: disable CentOS default repositories
This is a rework of 37b824c.

Centos 7:
Disable default repositories since the content of the ISO DVD is copied
in the os/ directory.

CentOS 8:
Instead of disabling the default OS repositories then configure new ones
in different files, use yum_repository parameter `file` to replace
existing configuration for BaseOS and AppStream.
Disable the Extras repository which is enabled by default but not
shipped with the ISO DVD. If one need to add the Extras repo, I would
recommend to add it to the inventory like any other repository.

Remove the now useless tunable `remove_distro_repositories`.