feat: split repository-root and images-root in pxe
The images used to boot (vmlinuz,initrd.img,squashfs) are located in
images-root which depends on a minor release of CentOS or RHEL. Those
images are provided in the ISO DVD.

The repository-root is the path to the packages which are available in a
`$releasever` directory, which allows to have a single stream for
updates, like CentOS and Red Hat provide.

This commit removes configuration_centos.ipxe.j2 which is unused.

PR: #149