feat: add support for slave DNS servers
The master DNS is the first element of the dns_ip in the primary
management network's configuration. I assume the primary management
network is the network of the Ansible controller.

  networks:
    ice1-1:
      services_ip:
        dns_ip:
          - 10.10.0.1  # Master DNS server
          - 10.10.0.2

The role dns_server will create the forward and reverse zones on the
master DNS server with the new task file master.yml. For each zone, the
role creates an SOA file that is included in the zone file. This allows
to update the zone serial only when the zone file changed. The serial
number is the Unix epoch (32-bits) at time of the zone change.

To add slave DNS servers, one need to configure the dns_slaves parameter
with a list of IP addresses in the primary management network. A good
practice would be to add this in the general_settings/network.yml
inventory file.

  dns_slaves:
    - 10.10.0.2
    - 10.10.0.3