feat: merge equipment_profile with combine
Because hash_behavior=merge will be deprecated with Ansible 2.10, move to the
combine filter to set the equipment_profile.

The default values are still defined in `equipment_profile`. To override any
default parameter for a specific equipment profile, define a new hash
`equipment_profile_override` with the parameter in the file
inventory/group_vars/equipment_*/equipment_profile.yml:

.. code-block:: yaml

  equipment_profile_override:
    firewall: false

We need to call the combine filter explicitely, while hash_behavior was
transparent, thus the new role `common`.