fix: define dnf variables in disklessset
It is possible to define dnf variables $environment and $releasever in
the repositories_client role, to set the baseurl of a repository. When
disklessset run dnf with a new installroot, dnf is not able to get the
values of those variables and eventually fails.

Since the script relies on the host configuration to get repos
information, make the script a jinja2 template file and define the
$environment and $releasever in the installroot from the host inventory.

If repositories_releasever is not defined, resolve to the releasever of
the host.