ci: check tags of Gather OS Variables tasks
The script lookup the variables in vars/main.yml, then search if these
variables are used in a template or in a task file. If so, it will check
if the task has a tag and if this tag is assigned to the first
include_var.

This ensures any call to execute a role/task with a tag will include the
requested variables first.

Remove Ubuntu_18.yml file which is non-YAML.

(cherry picked from commit 2f3fd57c7ed9a989ff8c53258b62e897f870f614)