feat: template custom NHC files from the inventory
This change allows to define the name and content of each custom files
one want to install in /etc/nhc/ in the user inventory. This helps to
keep all the configuration in the inventory instead of splitting some
values in the inventory for templates and in the role for files.