feat: assign a network interface to firewall zone
If you want to assign a network interface to a network zone, define
the zone in network_interfaces. The zone name will be added to the
ZONE parameter of the ifcfg file.

As for other changes made by the nic role, the network service needs
to be restarted for changes to take effect.