fix: skip ansible-lint rule 208 when recurse: yes
Remove mode: 0750 for file tasks where recurse: yes.

All the files in the directory will inherit this mode, which is wrong.
This mode was added in commit d70ad62 to resolve an error reported by
ansible-lint. Use #noqa 208 instead to skip the qa for this task, we can
rely on default mode set during the installation of the RPM.