Add high_availability_autostart variable
Adding high_availability_autostart so the user can control this behavior.

Also defining default behavior to be not enabling the HA resources (corosync & pacemaker) automatically.

I think that maybe we should not want the default behavior of the HA of the node is to start automatically on boot, because if we have some specific issue with the node that is leading to fence, It may get into some kind of fence-loop ( depending on the stickness/score configurations of the related resources as well ). So the main behavior maybe should be: "error"-> fence -> some kind of support evaluation -> bring node back to cluster. 

I may be wrong about this default behavior, just starting some topics, the variable I think we should have anyways thogh.