ci: update to ansible-lint v6.5
The recent release of ansible-lint v6.5 brings some breaking changes.
Pin the version to 6.5.x, so any future update won't break the CI out of
the blue.

Disable jinja[spacing]: there might be an interest to run it later.

Disable name[casing]: this is not compatible with the task naming
convention of this project.