clustershell upgrade: 1.2.1
(fix): Configured local.cfg template to ignore the node called dummy,
which was impacting generation of the @all nodeset. Fix bug #741

(feat): added feature so that the user can configure a set of custom mapping nodes and use the Clusters functionality (@nodeset:groups).

(docs): added tutorial on how to configure a custom mapping nodeset and how to use Clusters functionality with ClusterShell.